# Implement texture mipmap generation via compute pipelines