--- a/osu.Framework/Graphics/OpenGL/GLRenderer.cs
+++ b/osu.Framework/Graphics/OpenGL/GLRenderer.cs
@@ -275,6 +275,8 @@ namespace osu.Framework.Graphics.OpenGL
             int width = texture.Width;
             int height = texture.Height;
 
+            GL.UseProgram(computeMipmapShader!.Value);
+
             for (int level = 1; level < IRenderer.MAX_MIPMAP_LEVELS + 1 || (width > 1 || height > 1); level++)
             {
                 width = MathUtils.DivideRoundUp(width, 2);