… osu.Framework/Graphics/OpenGL/Shaders/GLUniformBlock.cs
… osu.Framework/Graphics/OpenGL/Textures/GLTexture.cs
… osu.Framework/Graphics/Rendering/ComputeMipmapGenerationParameters.cs
… osu.Framework/Graphics/Rendering/Dummy/DummyNativeTexture.cs
… osu.Framework/Graphics/Rendering/INativeTexture.cs
… osu.Framework/Graphics/Rendering/Renderer.cs
… osu.Framework/Graphics/Textures/TextureAtlas.cs
… osu.Framework/Graphics/Textures/TextureStore.cs
… osu.Framework/Graphics/Veldrid/Textures/VeldridTexture.cs
… osu.Framework/Graphics/Veldrid/Textures/VeldridTextureResources.cs
… osu.Framework/Graphics/Veldrid/Textures/VeldridVideoTexture.cs
… osu.Framework/Graphics/Veldrid/VeldridRenderer.cs
… osu.Framework/Platform/Windows/WindowsGLRenderer.cs
… osu.Framework/Resources/Shaders/sh_mipmap.comp