- [ ] Depends on https://github.com/mellinoe/veldrid/pull/506
- [ ] Depends on https://github.com/ppy/veldrid/pull/15
- Should (resolve) https://github.com/ppy/osu/issues/23252
- Should (resolve) https://github.com/ppy/osu-framework/issues/5748

## Preface

Since https://github.com/ppy/osu-framework/pull/5508, custom mipmap generation was implemented to reduce the overhead in the case of uploading small textures to an atlas.

The implementation revolved around binding one level of a texture as a sampler and another level as a render target (FBO), and using `texture` to downsample the larger texture and store the result in the smaller texture.

This was pretty straightforward for OpenGL and Metal, but D3D11 and Vulkan weren't very supportive. Both of them don't seem to like the idea of binding the same texture as both an input and output of the shader.

That was worked around by allocating another texture to bind as the input for the shader, but that required preparing the texture content for each level, which means switching between encoders on Metal (and Veldrid's Vulkan implementation [doesn't handle that well?](https://github.com/ppy/osu-framework/issues/5748)).

## Compute-based mipmap generation

After exploring through compute shaders for the past week, I've came up with a simple implementation that avoids creating extra resources as much as possible.

The concept of it boils down to a compute shader accepting an input texture, a linear sampler, and a uniform buffer supplying the uploaded region. The compute shader downsamples from the texture according to the specified uploaded region and the current invocation ID, and stores the result in the output texture.

For Veldrid, this is implemented by splitting each texture into texture views in `VeldridTextureResources`, bind one view for sampling and another for output, and supply a re-usable uniform buffer with region data. 

For Legacy OpenGL, this is implemented similarly, but texture views aren't required as each texture resembles a sampler by itself, so the properties can be modified accordingly for each dispatch call without creating extra resources.

Compute shaders is shown to be supported by 80% of systems [according to this game](https://feedback.wildfiregames.com/gl/extensions/), but the rendering-based path still exists as a fall back for older hardware since it still behaves 10x better than driver implementation. It can be further improved by using texture views and caching VBOs, but I'm leaving it as-is in this PR.

## Better rectangle merging implementation

After writing a test scene for mipmap generation, it turns out there was a high overhead coming from the rectangle merging logic added in https://github.com/ppy/osu-framework/pull/5508.

The loop works well for a few overlapping regions, but can get very complex when queuing a high number of uploads at different regions.

The overhead can get as high as ~500ms, which is 100x higher than the overhead of mipmap generation itself.

After [discussing this on discord](https://discord.com/channels/188630481301012481/589331078574112768/1099936570796154971) with @smoogipoo, I've came up with a different implementation that follows the nature of texture atlases and produces as less rectangles as possible to simplify the mipmap generation process.

The implementation accepts all uploaded regions, and produces one or two rectangles, one covering all uploaded regions that are horizontally after the top-left uploaded region, and another rectangle covering all uploaded regions that are horizontally before the top-left uploaded region.

| Before | After |
|--------|-------|
| ![CleanShot 2023-04-26 at 16 04 00](https://user-images.githubusercontent.com/22781491/234583919-bb207d22-059e-40fa-a414-65aee2c94806.png) | ![CleanShot 2023-04-26 at 15 54 05](https://user-images.githubusercontent.com/22781491/234583980-b95816ca-1015-4ca1-813c-2964e04b817b.png) |

On master, mipmaps would be generated for every single uploaded data (including the paddings around the texture) as can be seen above, this resulted in too many vertices and unnecessary fills.

With the new implementation, all uploaded regions are formed into two rectangles to dispatch as many threads as possible without overlapping / unnecessary work.

The new implementation is also ~75x faster than the original logic, which should reduce stutters in osu! when starting gameplay, as there are a lot of textures getting added to the atlas at once.

---

As an aside, I have noticed both the computing and rendering implementations missing certain textures when testing with osu!, I'll leave this PR blocked from merge while I look into it.

Requires testing on all configurations below:
 - [ ] Windows (Legacy OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader`)
   - [ ] Rendering path
 - [ ] Windows (Veldrid OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader` and `GL_ARB_texture_view`)
   - [ ] Rendering path
 - [ ] Windows (Direct3D 11)
 - [ ] macOS (x64, Legacy OpenGL)
 - [ ] macOS (x64, Veldrid OpenGL)
 - [ ] macOS (x64, Metal)
 - [ ] macOS (M1, Legacy OpenGL)
 - [ ] macOS (M1, Veldrid OpenGL)
 - [ ] macOS (M1, Metal)
 - [ ] Linux (Legacy OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader`)
   - [ ] Rendering path
 - [ ] Linux (Veldrid OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader` and `GL_ARB_texture_view`)
   - [ ] Rendering path
 - [ ] Android (Legacy OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader`)
   - [ ] Rendering path
 - [ ] iOS (Legacy OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader`)
   - [ ] Rendering path
 - [ ] iOS (Veldrid OpenGL)
   - [ ] Computing path (requires `GL_ARB_compute_shader` and `GL_ARB_texture_view`)
   - [ ] Rendering path
 - [ ] iOS (Metal)

Optional configurations:
 - [ ] Windows (Vulkan) 
 - [ ] Linux (Vulkan)